Mudando o atributo avaliaçao pra double

+++ src/minweb/modelo/Filme.java
@@ -25,7 +25,7 @@ public class Filme extends ObjetoBD {
 	@ElementCollection
 	private List<String> diretores;
 	@Basic(optional=false)
-	private int avaliação;
+	private double avaliação;
 	@Basic
 	private int ano;
 	@Basic
@@ -55,7 +55,7 @@ public class Filme extends ObjetoBD {
 	public List<String> getDiretores() {
 		return diretores;
 	}
-	public int getAvaliação() {
+	public double getAvaliação() {
 		return avaliação;
 	}
 	public int getAno() {
@@ -88,7 +88,7 @@ public class Filme extends ObjetoBD {
 	public void setDiretores(List<String> diretores) {
 		this.diretores = diretores;
 	}
-	public void setAvaliação(int avaliação) {
+	public void setAvaliação(double avaliação) {
 		this.avaliação = avaliação;
 	}
 	public void setAno(int ano) {
